B-1 refactoring to common library.
