+++ src/main/java/net/prottonne/lab/auth/service/JwtService.java
@@ -9,11 +9,11 @@ import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.Date;
 import java.util.UUID;
-import net.prottonne.lab.auth.constant.ErrorMessage;
 import net.prottonne.lab.auth.iservice.IJwt;
 import net.prottonne.lab.auth.entity.RequestLogin;
 import net.prottonne.lab.auth.entity.ResponseLogin;
 import net.prottonne.lab.common.util.auth.JwtUtil;
+import net.prottonne.lab.common.util.exception.ErrorMessage;
 import net.prottonne.lab.common.util.exception.TechnicalException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
